Fix event NPC issues

+++ Game.cs
@@ -39,9 +39,9 @@ namespace ReAction
         public static bool IsActionOutOfRange(uint actionID, GameObject* o) => DalamudApi.ClientState.LocalPlayer is { } p && o != null
             && getActionOutOfRangeOrLoS(actionID, (GameObject*)p.Address, o) is 566;
 
-        public delegate byte UseActionDelegate(ActionManager* actionManager, uint actionType, uint actionID, uint targetObjectID, uint param, uint useType, int pvp, IntPtr a8);
+        public delegate byte UseActionDelegate(ActionManager* actionManager, uint actionType, uint actionID, long targetObjectID, uint param, uint useType, int pvp, IntPtr a8);
         public static Hook<UseActionDelegate> UseActionHook;
-        private static byte UseActionDetour(ActionManager* actionManager, uint actionType, uint actionID, uint targetObjectID, uint param, uint useType, int pvp, IntPtr a8)
+        private static byte UseActionDetour(ActionManager* actionManager, uint actionType, uint actionID, long targetObjectID, uint param, uint useType, int pvp, IntPtr a8)
             => ActionStackManager.OnUseAction(actionManager, actionType, actionID, targetObjectID, param, useType, pvp, a8);
 
         public static void Initialize()
